Критические исправления системы категоризации
Исправлены ошибки:
1. Удалены импорты несуществующей функции get_last_expense_by_description
   - Удалён блок переиспользования суммы (строки 500-539)
   - Удалён блок проверки истории перед AI (строки 640-665)
   - Теперь пользователь должен всегда указывать сумму

2. Изменён usage_count с IntegerField на FloatField
   - Было: дробные значения (0.5) округлялись до 0
   - Стало: сохраняются дробные значения корректно
   - AI-категоризация: usage_count = 0.5
   - Ручное изменение: usage_count = 1.0
   - Миграция: 0049_alter_categorykeyword_usage_count_and_more.py

Теперь приоритет работает правильно:
- Ручные изменения (1.0) > AI (0.5)
- Очистка не удаляет свежие AI-слова (usage_count >= 0.5)

+++ expenses/models.py
@@ -378,8 +378,8 @@ class CategoryKeyword(models.Model):
         verbose_name='Язык ключевого слова'
     )
     
-    # Счетчик использования (только ручные исправления)
-    usage_count = models.IntegerField(default=0, verbose_name='Количество использований')
+    # Счетчик использования (дробные значения: AI = 0.5, ручное = 1.0)
+    usage_count = models.FloatField(default=0.0, verbose_name='Количество использований')
     
     # Нормализованный вес для конфликтующих слов
     normalized_weight = models.FloatField(default=1.0, verbose_name='Нормализованный вес')
@@ -1047,8 +1047,8 @@ class IncomeCategoryKeyword(models.Model):
     category = models.ForeignKey(IncomeCategory, on_delete=models.CASCADE, related_name='keywords')
     keyword = models.CharField(max_length=100, db_index=True)
     
-    # Счетчик использования (только ручные исправления)
-    usage_count = models.IntegerField(default=0, verbose_name='Количество использований')
+    # Счетчик использования (дробные значения: AI = 0.5, ручное = 1.0)
+    usage_count = models.FloatField(default=0.0, verbose_name='Количество использований')
     
     # Нормализованный вес для конфликтующих слов
     normalized_weight = models.FloatField(default=1.0, verbose_name='Нормализованный вес')
